move setParser to ModelManager
