+++ src/main/java/seedu/address/logic/commands/ListDeletedCommand.java
@@ -4,6 +4,7 @@ import static java.util.Objects.requireNonNull;
 import static seedu.address.model.Model.PREDICATE_SHOW_ALL_SOURCES;
 
 import seedu.address.logic.CommandHistory;
+import seedu.address.logic.LogicManager;
 import seedu.address.model.Model;
 
 /**
