+++ src/main/java/seedu/address/model/ParserMode.java
@@ -1,4 +1,4 @@
-package seedu.address.logic;
+package seedu.address.model;
 
 public enum ParserMode {
     RECYCLE_BIN,
